back-server/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==38.0.3
debugpy==1.6.2
decorator==5.0.7
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-cors-headers==3.13.0
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
entrypoints==0.4
executing==0.9.1
Faker==8.1.1
fastjsonschema==2.16.1
idna==3.4
inflection==0.5.1
ipykernel==6.15.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.1.2
jsonschema==4.17.0
jupyter-client==7.3.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyterlab-pygments==0.2.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.18
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.8.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2022.6
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.3.0
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.2
traitlets==5.0.5
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
